+++ core/ProfilerTools.cc
@@ -896,8 +896,8 @@ namespace Profiler_Tools {
 
       // number of lines and functions
       if (p_flags & STATS_NUMBER_OF_LINES) {
-        line_func_count_str = mputprintf(line_func_count_str, "%s:\t%zu lines,\t%zu functions\n",
-           p_db[i].filename, p_db[i].lines.size(), p_db[i].functions.size());
+        line_func_count_str = mputprintf(line_func_count_str, "%s:\t%lu lines,\t%lu functions\n",
+           p_db[i].filename, (unsigned long)(p_db[i].lines.size()), (unsigned long)(p_db[i].functions.size()));
       }
       total_code_lines += p_db[i].lines.size();
       total_functions += p_db[i].functions.size();
@@ -905,7 +905,7 @@ namespace Profiler_Tools {
     if (p_flags & STATS_NUMBER_OF_LINES) {
       line_func_count_str = mputprintf(line_func_count_str,
         "--------------------------------------\n"
-        "Total:\t%zu lines,\t%zu functions\n", total_code_lines, total_functions);
+        "Total:\t%lu lines,\t%lu functions\n", (unsigned long)total_code_lines, (unsigned long)total_functions);
     }
 
     if (p_flags & (STATS_TOP10_ALL_DATA | STATS_ALL_DATA_SORTED)) {
